Shift extensions to new file
