+++ Example/BlockV/Controllers/Actions/OutboundActionViewController.swift
@@ -134,58 +134,3 @@ class OutboundActionViewController: UIViewController {
     }
     
 }
-
-/// Action Extensions
-extension VatomModel {
-    
-    /// Clones this vAtom to the specified token.
-    ///
-    /// Note: Calling this action will trigger the action associated with this vAtom's
-    /// template. If an action has not been configured, an error will be generated.
-    ///
-    /// - Parameters:
-    ///   - token: Standard UserToken (Phone, Email, or User ID)
-    ///   - completion: The completion handler to call when the action is completed.
-    ///                 This handler is executed on the main queue.
-    public func clone(toToken token: UserToken,
-                         completion: @escaping ([String: Any]?, BVError?) -> Void) {
-        
-        let body = [
-            "this.id": self.id,
-            "new.owner.\(token.type.rawValue)": token.value
-        ]
-        
-        // perform the action
-        BLOCKv.performAction(name: "Clone", payload: body) { (json, error) in
-            //TODO: should it be weak self?
-            completion(json, error)
-        }
-        
-    }
-    
-    /// Redeems this vAtom to the specified token.
-    ///
-    /// Note: Calling this action will trigger the action associated with this vAtom's
-    /// template. If an action has not been configured, an error will be generated.
-    ///
-    /// - Parameters:
-    ///   - token: Standard UserToken (Phone, Email, or User ID)
-    ///   - completion: The completion handler to call when the action is completed.
-    ///                 This handler is executed on the main queue.
-    public func redeem(toToken token: UserToken,
-                      completion: @escaping ([String: Any]?, BVError?) -> Void) {
-        
-        let body = [
-            "this.id": self.id,
-            "new.owner.\(token.type.rawValue)": token.value
-        ]
-        
-        // perform the action
-        BLOCKv.performAction(name: "Redeem", payload: body) { (json, error) in
-            //TODO: should it be weak self?
-            completion(json, error)
-        }
-        
-    }
-    
-}
